Split class out
