+++ osu.Game/Online/Multiplayer/GameTypes/GameTypeTagTeam.cs
@@ -1,0 +1,41 @@
+using osu.Framework.Graphics;
+using osu.Framework.Graphics.Containers;
+using osu.Game.Graphics;
+using osuTK;
+
+namespace osu.Game.Online.Multiplayer.GameTypes
+{
+    public class GameTypeTagTeam : GameType
+    {
+        public override string Name => "Tag Team";
+
+        public override Drawable GetIcon(OsuColour colours, float size)
+        {
+            return new FillFlowContainer
+            {
+                Anchor = Anchor.Centre,
+                Origin = Anchor.Centre,
+                AutoSizeAxes = Axes.Both,
+                Direction = FillDirection.Horizontal,
+                Spacing = new Vector2(2f),
+                Children = new[]
+                {
+                    new SpriteIcon
+                    {
+                        Icon = FontAwesome.fa_refresh,
+                        Size = new Vector2(size * 0.75f),
+                        Colour = colours.Blue,
+                        Shadow = false,
+                    },
+                    new SpriteIcon
+                    {
+                        Icon = FontAwesome.fa_refresh,
+                        Size = new Vector2(size * 0.75f),
+                        Colour = colours.Pink,
+                        Shadow = false,
+                    },
+                },
+            };
+        }
+    }
+}
