+++ osu.Game/Online/Multiplayer/VersusRow.cs
@@ -1,0 +1,52 @@
+using osu.Framework.Graphics;
+using osu.Framework.Graphics.Containers;
+using osu.Framework.Graphics.Shapes;
+using osuTK;
+using osuTK.Graphics;
+
+namespace osu.Game.Online.Multiplayer
+{
+    public class VersusRow : FillFlowContainer
+    {
+        public VersusRow(Color4 first, Color4 second, float size)
+        {
+            var triangleSize = new Vector2(size);
+            AutoSizeAxes = Axes.Both;
+            Spacing = new Vector2(2f, 0f);
+
+            Children = new[]
+            {
+                new Container
+                {
+                    Size = triangleSize,
+                    Colour = first,
+                    Children = new[]
+                    {
+                        new EquilateralTriangle
+                        {
+                            Origin = Anchor.BottomLeft,
+                            RelativeSizeAxes = Axes.Both,
+                            Rotation = 90,
+                            EdgeSmoothness = new Vector2(1f),
+                        },
+                    },
+                },
+                new Container
+                {
+                    Size = triangleSize,
+                    Colour = second,
+                    Children = new[]
+                    {
+                        new EquilateralTriangle
+                        {
+                            Anchor = Anchor.BottomLeft,
+                            RelativeSizeAxes = Axes.Both,
+                            Rotation = -90,
+                            EdgeSmoothness = new Vector2(1f),
+                        },
+                    },
+                },
+            };
+        }
+    }
+}
