+++ osu.Game/Screens/Multi/Match/Components/GameTypePicker.cs
@@ -9,6 +9,7 @@ using osu.Framework.Graphics.UserInterface;
 using osu.Framework.Input.Events;
 using osu.Game.Graphics;
 using osu.Game.Online.Multiplayer;
+using osu.Game.Online.Multiplayer.GameTypes;
 using osu.Game.Screens.Multi.Components;
 using osuTK;
